librados: add async interfaces for use with boost::asio
adds librados interfaces that conform to the Extensible Asynchronous
Model [1] of boost::asio. this model allows the last 'handler' argument
to either be a callback that gets the result, a coroutine yield_context
that will suspend until completion, or 'use_future' to request the
result in a std::future. see the unit test for examples of each

these new interfaces are called 'initiating functions' in asio, and they
come with some very specific requirements [2]. while most of those
requirements are met here, some exceptions are made to avoid abi changes:

- instead of making the 'associated io_service object' available through
  a get_io_service() member function on the first IoCtx argument, it is
  passed directly as the first argument
- several arguments are forwarded directly to the IoCtx member function
  that they wrap. the types of these arguments are preserved, instead of
  changing the const qualifier, reference or pointer type to match their
  lifetime requirements

[1] "Library Foundations for Asynchronous Operations"
http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2014/n3896.pdf
[2] "Requirements on asynchronous operations"
http://www.boost.org/doc/libs/1_65_1/doc/html/boost_asio/reference/asynchronous_operations.html

Signed-off-by: Casey Bodley <cbodley@redhat.com>