qa/distros/podman: replace centos 8 with centos 9
This commit is based on https://github.com/ceph/ceph/commit/c8873c6591d368e12907669c274fd3d6391e3f68,
although it is a direct merge to quincy due to backport complexities.

We needed the container-hosts version of podman on centos 8,
but there's no need to keep using it in centos 9

Additionally, changes from crun variants to runc variants since
crun is now the default container runtime on centos 9.

Centos 8 is end of life, so we must use centos 9.

Signed-off-by: Laura Flores <lflores@ibm.com>