osdc: enum OpCancellation and "safe" cancellation
define a "safe" type of cancellation that only takes effect if/when
Objecter can guarantee that the operation had no side effects

read operations never have side effects, so safe cancellation behaves as
normal. write operations can be safely cancelled unless there's one or
more in-flight MOSDOp requests that may still succeed

once safe cancellation is requested, no further redirects or retries are
attempted. if we get a successful reply, the op still succeeds.
otherwise, the operation is cancelled if/when we receive a reply to the
final outstanding request

Signed-off-by: Casey Bodley <cbodley@redhat.com>