rgw/rados: preserve prefetch_data over follow_olh()
RGWObjectCtx tracks a separate RGWObjState for each object name.
Object::set_prefetch_data() only sets that prefetch flag for the
head object name. but in the case of versioned objects, that name
corresponds to the object logical head (olh) object. when follow_olh()
forwards us to the target head object (the current version), our
call to get_obj_state() no longer sees the prefetch flag

as a result, the prefetch optimization for GetObject doesn't work for
versioned objects, so we have to send a second rados request to get its
data

Signed-off-by: Casey Bodley <cbodley@redhat.com>